+++ lib/zenlish/wclasses/auxiliary.rb
@@ -2,9 +2,10 @@ require_relative 'verb'
 
 module Zenlish
   module WClasses
-    # The verbs: be, have, do
+    # The verbs: be, have, do and the modal verbs can, could, may, must, shall,
+    #   should, will, would
     # As they are so important to the English language, they deserve
-    # their own class
+    # their own class(es)
     class Auxiliary < Verb
     end # class
   end # module