+++ spec/zenlish/wclasses/verb_spec.rb
@@ -5,6 +5,13 @@ require_relative '../../../lib/zenlish/wclasses/verb' # Load the class under tes
 
 module Zenlish
   module WClasses
+    class Verb
+      # Do limited monkeypatching, just for testing purposes
+      def init_paradigms
+        # Empty method
+      end
+    end
+
     describe Verb do
       subject { Verb.new }
 
@@ -18,11 +25,11 @@ module Zenlish
         it 'should know that it has inflected forms' do
           expect(subject).not_to be_invariable
         end
-        
+
         it 'should know its feature definitions' do
           expect(subject['NUMBER']).to be_kind_of(Feature::FeatureDef)
           expect(subject['PERSON']).to be_kind_of(Feature::FeatureDef)
-          expect(subject['PARADIGM'].default.val).to eq('Regular_inflection')           
+          expect(subject['PARADIGM'].default.val).to eq('Regular_inflection')
         end
       end # context
     end # describe